refactor: move fetch request logic to load function
